[Store] Fix 32-bit host-pointer offset overflow in SPDK NoF transfers over 4 GiB
The SPDK NoF worker computes the host buffer offset for each sub-I/O as
`lba_off * block_size`, where lba_off was an int and block_size a
uint32_t, so the product is evaluated in 32-bit unsigned arithmetic.
Once the running offset reaches 4 GiB (lba_off * block_size >= 2^32) it
wraps around and submit_ptr addresses the wrong part of the buffer: the
DMA for the tail of a >4 GiB transfer then reads from / writes to the
start of the user buffer instead, silently corrupting data. submit_lba
two lines below is already widened to 64-bit, so only the host pointer
offset was left narrow.

A single SpdkNofTask carries lba_count = size / block_size for the whole
request with no size cap, so any NoF transfer larger than 4 GiB reaches
this path (the same >4 GiB scenario fixed for SSD offload in #2570).
Widen lba_off to uint64_t so the offset arithmetic stays 64-bit.

+++ mooncake-store/src/transfer_task.cpp
@@ -485,7 +485,13 @@ void SpdkNofWorkerPool::workerThread(int work_idx) {
                         uint32_t submit_lba_count = std::min(
                             avail_blocks, std::min(task->remaining_lba,
                                                    nof_qos->blocks_per_chunk));
-                        int lba_off = task->lba_count - task->remaining_lba;
+                        // 64-bit: a NoF transfer can exceed 4 GiB, and once
+                        // lba_off * block_size crosses 2^32 the host-pointer
+                        // offset would wrap when computed in 32 bits (int *
+                        // uint32_t), making submit_ptr address the wrong part
+                        // of the buffer. submit_lba below is already widened.
+                        uint64_t lba_off =
+                            task->lba_count - task->remaining_lba;
                         uint64_t submit_lba = task->lba + lba_off;
                         void* submit_ptr = reinterpret_cast<void*>(
                             reinterpret_cast<char*>(task->ptr) +
